Move RestExportService to dataExport package
